+++ gui/src/components/mainInput/InputToolbar.tsx
@@ -5,8 +5,8 @@ import {
 import { PhotoIcon as SolidPhotoIcon } from "@heroicons/react/24/solid";
 import { InputModifiers } from "core";
 import { modelSupportsImages } from "core/llm/autodetect";
-import { useMemo, useRef, useState } from "react";
-import { useSelector } from "react-redux";
+import { useEffect, useMemo, useRef, useState } from "react";
+import { useDispatch, useSelector } from "react-redux";
 import styled from "styled-components";
 import {
   defaultBorderRadius,
@@ -26,6 +26,10 @@ import {
 } from "../../util";
 import ModelSelect from "../modelSelection/ModelSelect";
 import { isBareChatMode, isPerplexityMode } from '../../util/bareChatMode';
+import { setDefaultModel } from "../../redux/slices/stateSlice";
+import { RootState } from "@/redux/store";
+import { useLocation } from "react-router-dom";
+
 
 const StyledDiv = styled.div<{ isHidden: boolean }>`
   padding: 4px 0;
@@ -96,6 +100,28 @@ function InputToolbar(props: InputToolbarProps) {
   const perplexityMode = isPerplexityMode();
 
   const useActiveFile = useSelector(selectUseActiveFile);
+  const allModels = useSelector(
+    (state: RootState) => state.state.config.models,
+  );
+
+  const dispatch = useDispatch();
+  const location = useLocation();
+
+  useEffect(() => {
+    console.dir("IN NEW EFFECT")
+    console.dir(location?.pathname)
+    if (location.pathname === "/aiderMode") {
+      const aider = allModels.find(
+        (model) => model?.title?.toLowerCase().includes("aider"),
+      );
+      dispatch(setDefaultModel({ title: aider?.title }));
+    } else if (location.pathname === "/perplexityMode") {
+      const perplexity = allModels.find(
+        (model) => model?.title?.toLowerCase().includes("perplexity"),
+      );
+      dispatch(setDefaultModel({ title: perplexity?.title }));
+    }
+  }, [location, allModels]);
 
   return (
     <>
@@ -219,3 +245,4 @@ function InputToolbar(props: InputToolbarProps) {
 }
 
 export default InputToolbar;
+
